Drop nodescan worker from node when complete
Once we have completed the nodescan process, we can drop the reference
to the nodescan worker.  Keeping it was harmless, but it did mean that
we would attempt to stop it twice and emit a second log line when we
deleted the node.  This keeps things tidy and avoids that.

Change-Id: I4408114f2afed0281670b11d373eaa35923d0a27