DNM: Add replication delay to Gerrit
It may be possible for some Gerrit replication configurations not
to emit the replication events Zuul now expects.  This change is
intended as a temporary stop-gap measure while we work to understand
the behavior and develop a proper fix.  It restores something similar
to the previous 10-second permanent delay.

Set the replication_timeout to 0 when using this change.

Change-Id: Ibebe794d256895910bf672fe539cec2a0e17d986