Avoid duplicate launcher upload jobs
Image upload jobs can take some time, and during that time, the
launcher may be triggered to check for missing uploads multiple
times.  If that occurs, it may add many instances of UploadJob
to its queue for the same pending uploads.

This can waste resources as the queue grows ever larger.  With
the recent change to create new upload objects on failure, it will
also cause us to continually create those new objects for the same
failed uploads.  This can happen if a lot of upload jobs are queued
before the initial failure is detected; the remaining jobs in the
queue will all each create a new retry upload object.

To avoid this, we will keep track of which upload uuids we have
UploadJobs currently queued, and will avoid creating duplicates.

Change-Id: Ia0e7394b54aa2064093b48e998c5061d13601a31