Require multinode requests served from same provider
The zuul-launcher currently attempts to supply nodes from the same
provider if possible, but it can fall back to using multiple providers
if there are persistent errors.  OpenDev has found that such errors
happen frequently enough (likely due to cloud quota issues which are
not detectable by Zuul as quota issues) that the behavior is troublesome.

This change alters the behavior so that ther is an internal behavior
switch: if it is at all possible for a node request to be fulfilled by
a single provider, then the launcher will insist that happen or the
request fail.  If it is not possible, then the current best-effort
behavior is retained.

Specifically the behavior is:

* If all labels can theoretically be supplied by a single provider,
  enable the new behavior.
* If the new behavior is enabled, then as soon as the first provider
  attempted reaches its max attempts and is removed from the candidate
  list, all nodes in the request will shift to a new provider.
* This may leave extra ready nodes assigned to the request (if a node
  has already succeeded on a provider while a second node fails) but
  as soon as the request is complete, the extra ready nodes will
  be deallocated and made available to other requests.
* If all providers fail, the request will fail with a node_error.

A future enhancement may deallocate the unused nodes sooner, but this
change does not (the current behavior is to avoid touching any unused
nodes during request processing, whether successful or not).

Change-Id: I6daa8a5450ed24d6b6ff736d25c0d68be924ecd4