Launcher: implement preferred provider
Zuul requests nodes from the same provider for children of paused
jobs.  Nodepool honors those requests, but zuul-launcher did not.
This change implements the same behavior in zuul-launcher.  It is
still a "best effort" -- if the launcher is unable to satisfy the
request from the preferred provider, it will use others.

Change-Id: Ic7ec01052c9dc4a6f2aad6263aec45ee612d2447