Launcher: skip non-pending uploads
If more than one launcher is running, they can race each other
in attempting to upload images.  They use locks to ensure that
only one processes an image at a time.  After acquiring the lock,
they check to see if the upload was already completed, and if so,
they skip it.  But they do not check whether the upload changed
state, for example, if it changed to the "deleted" state because
it failed and is being replaced.

Update the check so that we proceed with an upload attempt only
if our upload is still in the "pending" state after we lock it.
Otherwise, the main event loop will figure out what to do with
the upload next.

An explicit test for this is not added due to the complexity of
the race.

Change-Id: I865f07b2be93bd588a112c8a54cccd85ef5a6a2d