Speed up launcher tests and fix races
The launcher has a number of built-in delays.  Let's have the
unit test reduce those delays so that we spend less time sleeping
and more time testing.

This reduces the cumulative runtime for test_launcher.* tests by
almost half (327s vs 609s).

This addresses some race conditions in the tests.  It also makes
a change to the launcher itself to detect a condition where an
upload appeared in the cache while the launcher is deciding which
uploads to keep.  In that case, we will ignore newly created
uploads (which means we keep them).

Change-Id: I8953965f04a4578c312d02767ce0cb0a54e96c94