Improve watchdog logging
This adds some log messages that will allow us to better diagnose
potential problems with the executor watchdog.  We will get
start and stop messages with an identifier for the particular watchdog
timer (so we can associate it with its playbook).  The log messages
are also annotated with the build uuid.  If any exceptions happen in
the thread, they will be logged.

A misleading log message about the disk job killer is also removed
(it no longer is started or stopped; it simply runs all the time
now).

Change-Id: Ic1188b07b9ff465dd438010064e2245872e6998b