Disable leaked nodeset request cleanup in some tests
Some tests use a "fake" nodeset request in order to exercise the
launcher.  These requests will be unknown to the scheduler, therefore
it is unsafe for the scheduler to run the cleanup routine that deletes
leaked nodeset requests.

If we call the requestNodes method in a test (which generates these
fake requests), grab the node(set) request cleanup lock to prevent
the scheduler from running it.  There is no need to release the
lock.

Also, lock the leak detection test which does something similar.

Change-Id: Ieba1d5eda3c05d0c5cce5f08f6b9f0ca084fb45b