Launcher: fix error when retrying multinode request
When a multinode request has at least one node fail, we attempt
to replace that node with a new one.  However, an error caused us
to replace *all* the nodes in the request, while some nodes in
the request might be ready and still assigned to the request.  This
could leave a large number of ready+assigned nodes in the provider
and deadlock it due to quota starvation.

Correct the comparison and update a test to verify the behavior.

Change-Id: Ia4ec6fe3145eb7f8e44c67673ec931dfe4ec25f8