Remove python 2.7 stream testing
It is no longer possible to build the python2.7 container used for
stream testing.  Additionally, we are about to add support for
Ansible 11 which does not support python2.7 on target hosts.  We
are nearing the end of the end of python2.7, so we will accept the
risk of regressions and terminate testing.

Change-Id: If2ba9e2612201f712eb4b5f12b70dd5b7d5d40fd