Improve handling of unassigned ready nodes
In OpenDev, we noted a provider which was (due to other earlier
bugs) entirely full of unassigned ready nodes.  They are not
being used because, as a provider that is at quota, it has the
lowest priority when selecting providers.

This change:
* Attempts to better utilize ready nodes by sorting the provider
  list for a request by usable ready nodes so they are used sooner
* Includes usassigned ready nodes in the REST api output of every
  provider that can use them
* Does the same for reported stats
* Updates the _getProviders() method to avoid returning duplicate
  providers (which is harmless till now, but could end up
  counting unassigned ready nodes multiple times for the same
  providers).

Tests are added for the stats and REST api, but not the behavior
change for selectProviders.  This behavior is getting a little
too subtle to test in non-deterministic functional tests.  We may
need to invest in some unit tests of selectProviders.  For now,
we rely on the existing tests to at least avoid a regression.

Change-Id: If72a5a6c674db494a37b54828d024e2c7fe6a193