Reduce endpoint quota limit queries
Currently, after each reconfiguration, we call "postConfig" on every
provider to allow it to update any information necessary in the quota
cache.  We do a number of things to minimize API calls, including only
calling postConfig once for each provider on a launcher, and nominating
a single launcher to handle each provider.  However, since the limits
are a characteristic of an endpoint, we can still end up calling the
APIs more than once for each endpoint.

To further optimize this, we will split our work so that postConfig
is still called for every provider in order to update provider-specific
information, but we will refresh the quota limits only once per endpoint.
We switch the launcher nomination to be for the endpoint, so that only
the designated launcher will force updates on its designated endpoints
and any providers that reference that endpoint.

(As before, any launcher may still perform API queries if information
is missing from the quota cache.)

Change-Id: I5e3fb3bdfc19b8025346f339505039a295847df0