Fix test race in launcher test_jobs_executed
We assert that there are no nodes, requests, or associated locks
at the end of these two tests, however, it is possible for a lock
znode to be deleted just after the underlying ProviderNode znode
is deleted, in which case it would fail the assertion.

Replace the assertions with a loop that will allow a little grace
period for the lock to be deleted.  We apply this for all the
checks for simplicity even though it's not strictly necessary for
the provider nodes themselves.

Change-Id: I8661cc0ce39fb2dedd115c46a6401034c9d18cf8