Launcher: ignore failed nodes for request locality
When selecting providers for nodes, we may select the main provider
for a request based on the providers for existing nodes.  But we
shouldn't need to consider the provider for a failed node when doing
so.  By relaxing this restriction, this will allow the launcher to
select a different provider when retrying a node after a tempfail
(for quota).  If all of the nodes of a multinode request happen to
be in the failed state at the same time (unlikely!), the entire request
may shift providers.  Otherwise, we will keep the current behavior
of using the provider that the remaining nodes use.

But in the very common case of a single node tempfailing, we no longer
have any reason to retry on the same provider, so allow it to shift.

Change-Id: I2098bb5d830e8939d22def0331c527c1f1d4597e