Implement reject filter in GitLab driver
The pipeline configuration in Zuul allows specifying
what is required in order to enqueue changes, as well
as what are conditions to reject an item from the queue.
However, the rejection filtering was not implemented
in the GitLab driver so far. This commit introduces
the missing mechanism, based on the existing implementation
of the GitLab ref filter.

The `reject.<driver>.labels` attribute follows the logic
of `hashtag` attribute in Gerrit driver – if any specified
label will be in the Merge Request, it will be rejected.

Change-Id: I36cc5e0c13b9564a43875c733b0589c81ca94f5d