Add unit tests (and fixes) for selectProviders
This method is critical to launcher behavior, but can be somewhat
non-deterministic and has a number of pre-requisites, making it
difficult to test.

This change adds a unit test framework for this method so we can
evaluate its behavior deterministically based on a variety of input
scenarios.  We fake out the node cache, nodes, and requests; as well
as control the order it considers providers all in order to make its
behavior determisistic.

With this is place, it is easier to add test scenarios.

Several tests are combined into one in order to re-use the setup and
teardown (we still run a launcher even though it isn't doing
anything).

This work has uncovered two bugs that are fixed in this change:

* When deciding which providers an unassigned ready node can be used
  with, we need to conisder the endpoint, not just the connection, so
  that a node from region A isn't mistakenly handled through a
  provider for region B.

* Because we now remember which providers can work with an unassigned
  ready node in filterReadyNodes, we need to keep track of all of
  them, not just the first, so that they are all available to
  selectProviders.

Change-Id: I8c8a68f710246cb2278a3bcba01816d3581808a3