Launcher: handle delete state machine errors
If the delete state machine raised an exception, we would just try
to advance the state machine again, which may simply return the same
error (or in the case of the openstack delete state machine, raise
the same deferred threadpool exception again, resulting in an
ever-growing stacktrace).

Add a general exception handler for all delete state machines that
simply ends the state machine, allowing the znode to be deleted.  If
the node remains, the leaked node detection will handle it later.
This is similar to what happens if the delete state machine times out.

Change-Id: I6e4cdd7e20007c724a8fdae3fc755e51c93438d4