Fix test_max_ready_age
It was possible for this test to skip over the main part of the
test: verifying that an old ready node was deleted.  In that case,
it might be performing the act of replacement while the test ended,
which may produce an error.  Or it may have been able to complete
the test without error (but also not demonstrating the behavior).

Correct this by ensuring that the node we make "old" is actually
deleted, then wait for all the nodes (which should include its
replacement since that should happen in the same run loop) to
be ready.

Change-Id: I7d04248b6b818d0bd5807771054486ad36936b63