Fix test_launcher_image_expire_failed_upload
At one point during this test, we trigger a second artifact build,
but we don't wait for it to appear (instead, we can proceed after
only one artifact for the specified image is built).  This is racy,
since we are looking for exactly one artifact in order to proceed.

Correct this by waiting for the second artifact to be built before
proceeding.  Later in the test, we also wait for 2 artifacts; in
that case, we're expecting a third one to be created and the first
one to be deleted.  We can't catch the moment when there are 3
artifacts, but we can ensure that we run the deletion run, wait until
there are 2, then verify the oldest is not among them.

Also, this test had an extra check to wait for confing synchronization
which was not necessary.

Change-Id: I66bb9d914b75e454936cfffa756c0a5044d0f3c7