Allow control of upload methods
Users may want to disable certain upload methods (e.g., the "copy"
method) if that method doesn't work as well in their environment
as others.  Allow users to control that via the image configuration,
but default to allow everything.

Use this in some of our tests which would otherwise start copying
images.

Change-Id: I0dd1ec6c87944933b747045274b752ead458c058