Add snapshot-timeout label attribute
This allows users to set a timeout for how long a snapshot is
expected to take.  It is attached to the label of the node
that is being snapshotted since, to some extent, it is a characteristic
of the source node (though it also depends on what the job does to
it).  We don't know what image will ultimately be produced by the
snapshot (and it could be a 1:n relationship), so we don't attach
the setting it to the image.

Another possibility to consider would be to have the value returned
in the zuul return data.  That could be added in the future if desired,
with this value becoming a default.

Change-Id: I7c3bace87e966793d773171e61f4a6e1a1e08f50