Allow test override of launcher node ip
Some upcoming changes will be our first changes that will have
unit tests that run ansible on a fake test node provisioned by
zuul-launcher.  To allow tests to overide the IP address of the
returned node to localhost, update the AWS driver to build the
create instance arguments in a method that we can override in
unit tests.

Change-Id: I160b3a09620f2b386de659e975828481f396a44a