Replace zuul.node variable with zuul_node
The (not yet documented) job variable "zuul.node" doesn't actually
work because of the way we set zuul.* variables.  It is intended
as a replacement for the "nodepool" variable.  Instead, let's make
it "zuul_node" and set it just like we set "nodepool" today.

Also, add the node UUID to the values included.

Change-Id: I0bf25494b598e7fe5000be4dd94fa31121291522