Implement AWS snapshot import
This implements snapshot image handling in the AWS driver.  This
allows a user to snapshot a running Zuul node and that will be used
by zuul-launcher as an "artifact" and image "upload".  The snapshot
process returns an AWS ARN, which is a globally unique identifier.
That is used as the artifact URL.  The image "import" handler is
updated to recognize an ARN as a snapshot that can be directly
registered as an AMI.

The snapshot is tagged with an expiration time since, at the time
it is created, we don't know what upload uuids will be associated
with it.  But when we create the image, we re-tag the snapshot
with the upload uuid, effectively transferring ownership of it to
the image upload object.

Change-Id: Ia18d1a77d3b2d41a39f4dd4d46a5650112b4223c