web: Fix filtering on PipelineDetails page
Change 940500 introduced a bug that rendered the filtering functionality
on the PipelineDetails page unusable because a function argument was
passed to onFilterChanged where an array was expected.

This change fixes this and also simplifies the generation of filter
categories by doing this only once when the pipelines are loaded instead
of several times throughout page rendering. This fixes another bug where
select options (for the queue filter) could be incomplete because they
were generated based on pre-filtered pipeline states.

Also, an unrelated React warning about a wrong prop type in the
PauseModal is fixed.

Change-Id: I313bcdb430ed78252806204c4e261ab74c263dff