Correct example connection names in requirements docs
Generally we use conenction names with the same name as the
driver in our example documentation, as well as in the tutorial.
This generally makes sense for single-connection systems, or even
systems with only one connection for a given driver.

Some parts of the docs (the requirements filters) used a different
connection name with a "my-" prefix.  This change removes the
prefix in order to be consistent everywhere.

Change-Id: Ia98621cdfed7205232b90e16e8f93ffb2578cf0d