Limit upload retries to 3 attempts
This adds a limit to the number of attempts we will make to retry
a failed image upload.  Previously if an upload failed, we would
enter an infinite loop trying to upload it.  If the upload never
succeeds, this would either cause the launcher to consume bandwidth
downloading the image repeatedly, or, if the image did not need to
be downloaded, cause churn while we rapidly cycled through upload
zknodes.

Rather than do this, limit the number of attempts to 3, and then
mark the upload as failed.  The record will stay in ZK so that
the failure is visible, until it becomes older than ready uploads.

This should also fix test_launcher_image_expire_failed_upload, since
that test was racing the creation of upload objects.  This allows
the test to wait for the uploads to enter a steady state (either
ready or failed) before proceeding to the next step.

Change-Id: I4b4da3371a36e4b748acc3db4187bcb9e225d4c6