Add image, flavor, label defaults section
Many of the attributes of images, flavors, and labels are also
able to be set at the root of the provider for convenience (e.g.,
so that "host-key-checking" can be turned off for all labels, or
the "gp3" volume type can be set for all flavors).  But in some
cases, it can be ambiguous to actually use this feature.  Some
attributes are shared between images, labels, and flavors, and when
used with an image they will apply to the creation of the image
and when used with a label they will apply to the creation of the
instance.  There may be cases where a user would want to apply
a value to all labels, but not apply it to all images.

To make this more clear, this change new *-defaults top-level
keys instead; these will now be used for the inheritable attributes
for each of the three distinct objects.

To aid continuous deployment in OpenDev, backwards compatibility
is temporarily maintained.

Change-Id: I5f295f351b0feb2160f80fc613da1a1826691223