Further fix test_launcher_image_expire_failed_upload
This test is less racy than before, but there is currently a check
in it that can race:

1) We build an image and failed to upload it; this left a single
   upload for that image in a "failed" state.
2) We make a new build of an image with a successful upload; this
   addad an upload in the "ready" state.

A) At this point we checked for two uploads.

3) The launcher deleted the "failed" upload because it was older
   than the most recent "ready" upload.

That leaves only one upload for the image; if the check at A happens
after step 3, it will fail and timeout.

Instead, since it appears that we can reliably reach the steady state
at 3, we should move the check after it.

Change-Id: Ie108c32d44470e4d1a81913399f069ea985e0648