Fix test_nonexistent_lock and test_efficient_removal
These tests could attempt to grab an event from the merger queue
before the merger queue client object cache was updated with it.
Fix this by adding the event wait which is used in similar tests
(and evidently intended for this test since the event is present
and set but inadvertently unused).

Change-Id: Iee22281b764c39b0e65eed40dd3e4a660d284bba