Add AsList to AWS driver
Zuul configuration has a common pattern where lists of items may
be specified as a list or as a single item.  Update the AWS driver
so that the settings that can benefit from that use it.

To make this easier, we add a new voluptuous validator that takes
care of both parts of this (adjusting the schema and converting
the value so that it is always a list after validation).  This can
later replace both the to_list and as_list methods in configloader.

Change-Id: I7b139d7fc8fc54d4384bcf5f3c2b4a6e7404add8