AWS: Add security-group-ids parameter
This parameter is specified as a list when we interact with AWS,
but we only allow one value.  Update our schema to accept a list,
and rename the parameter in the process to match our pattern of
using a plural name when we accept a list.

The old name will be removed.

Change-Id: I1c7a728a62b84c53b4c984dd1bcfa4fe0e6f45cd