Avoid relative priority updates on completed requests
When we are updating nodeset requests to adjust their relative
priority, we obviously can not do so for completed requests.
Currently, we attempt to look up all requests in ZK and if they
are missing, we ignore the error and continue.

We can save the ZK request by checking whether we have an internal
record of the completed nodeset info before we attempt to lead
the request from ZK.  This will reduce ZK traffic and spurious
error messages in logs.

Change-Id: I6358b93b3f47e24b0e4d7cd5013f2496504829cd