Launcher: improve overfull allocation with incomplete quota info
When the launcher decides which provider to use for a request, it
sorts the providers by the percentage of quota used, and that can
be > 100% if providers have a backlog.   However, if we encounter
a provider with a quota attribute of 0, we will cap the result at
100%.  This could mean that the provider might appear to be the
east backlogged even if it is not.

To reduce the chances of this, continue to evaluate all of the quota
attributes, even if one is 0.  For the one that is zero, treat it
as 100%.  So if we have a quota of 100 instances, 0 cores, and a
backlog of requests for 10 more instances, the quota utilization will
be the expected 110% instead of 100%.

Also, add an additional log line that can help debug behavior
related to ready nodes.

Change-Id: I1d71fce5a6c49c22a632623927e41d9da2cafd91