Fix race with node provisioning
If the launchers are very slow, we can send the provisioned event
and the schedulers can even process it before we actually update
the request state to "fulfilled".  That will cause the schedulers
to think the request has failed.

Correct this by setting the state first, then sending the event.

Change-Id: I794cd704eb3e4d844dc5e3b8ed7320abadda09e1