Include requested nodes in quota percentage calculation
When we decide the best provider to use for a request, we order
the list of providers by their quota usage.  This only includes
actual usage, not "requested" nodes (any nodes that we have
already decided should be filled by that provider but haven't
started building yet).  If we end up with any significant number
of requested nodes, this could mean that we keep choosing one
a provider that has some available quota even if its backlog
of requested nodes is higher than other providers.

One way to end up with a bunch of requested nodes for a provider
is due to multi-node requests, which will cause multiple requests
for the same provider even if only one slot is available.

To address this, include the requested nodes when considering the
resource consumption of a provider.  This will sort the providers
by least backlog, which should produce a much more even distribution
of requests, allowing requests to find providers with low usage and
backlog, and not be tricked into selecting providers with, perhaps,
capacity for one node but a huge backlog.

We still need the quota used excluding requested nodes for stats,
so we now keep track of both totals in the node cache.

Finally, the quota objects in the node cache could grow without
bound since we stored the None object.  Correct that by removing
the quota object when the corresponding node is deleted.

Change-Id: Id1497637b91bf62c95da01ec239b3a486458c12a