Omnibus change to build image
This is to generate an image build we can use in opendev
to address the issues fixed by the following changes:

Include requested nodes in quota percentage calculation

When we decide the best provider to use for a request, we order
the list of providers by their quota usage.  This only includes
actual usage, not "requested" nodes (any nodes that we have
already decided should be filled by that provider but haven't
started building yet).  If we end up with any significant number
of requested nodes, this could mean that we keep choosing one
a provider that has some available quota even if its backlog
of requested nodes is higher than other providers.

One way to end up with a bunch of requested nodes for a provider
is due to multi-node requests, which will cause multiple requests
for the same provider even if only one slot is available.

To address this, include the requested nodes when considering the
resource consumption of a provider.  This will sort the providers
by least backlog, which should produce a much more even distribution
of requests, allowing requests to find providers with low usage and
backlog, and not be tricked into selecting providers with, perhaps,
capacity for one node but a huge backlog.

We still need the quota used excluding requested nodes for stats,
so we now keep track of both totals in the node cache.

Finally, the quota objects in the node cache could grow without
bound since we stored the None object.  Correct that by removing
the quota object when the corresponding node is deleted.

Original-Change-Id: Id1497637b91bf62c95da01ec239b3a486458c12a

Sort nodes by request time

Whenever a node launch fails and is retried (which may happen often
due to quota), a new node record is created with a random uuid.

When we iterate over nodes to process them, we use the order that
they are returned from the cache.  This may be chronological order
if the cache has been sucessfully watching new nodes being created,
or it may be alphabetical if the launcher recently started up and
read the cache from zk.  Or it may be a mixture of the two.

This means that, at best, the order that nodes are created when there
is a backlog of nodes in the "requested" state is random.  At worst,
it's the worst case where older nodes are disadvantaged by being
replaced at the end of the list when failing.

(Note that the nodeset requests are sorted by priority and time, but
once a node is actually created with the intent to build an instance
in the cloud -- this is what is currently un-ordered.  This is not
a concern unless providers are backlogged with nodes allocated to
them in the "requested" state.)

To address this, store the time of the original nodeset request on
each of the nodes created for it (or, the current time in the case
of min-ready).  Sort nodes by request time when processing them,
so that we always attempt to drive the oldest ones first.

Existing nodes in ZK will have a request time of 0, so they will
all appear equally very old on upgrade.

The output of the nodes REST api endpoint will now also be sorted
by the same metrics.

Original-Change-Id: I945ab9719f6904c0b0c2d0048cfcd8bbbb8acfde

Fix race with node provisioning

If the launchers are very slow, we can send the provisioned event
and the schedulers can even process it before we actually update
the request state to "fulfilled".  That will cause the schedulers
to think the request has failed.

Correct this by setting the state first, then sending the event.

Original-Change-Id: I794cd704eb3e4d844dc5e3b8ed7320abadda09e1

Always send nodes provisioned event before unlock

Make sure we always send the nodes provisioned event before unlocking a
node. In some edge-cases this could lead to duplicate events, which the
scheduler can already handle.

Original-Change-Id: I9cced63e3a55d63c3095d06bad67eabdd804f40c

Ensure node request event is sent after crash

We recently changed the order of events in the launcher so that
the node request complete event is sent after the state attribute
is set.  That is so that the scheduler can be assured that all
the attributes it needs are correct when it receives the event.
But it also means that if we crash between setting the state on
the request and sending the event, there will be no indication
to a subsequent launcher that picks up the request that the event
still needs to be sent.  It would get stuck until someone manually
dequeued a change.

To address this, we add another attribute to the request to
indicate that the event has been sent.  The intended audience is
only the launcher, not the scheduler.  Subsequent launchers can
see that a "fulfilled" request with an event_state that is not
COMPLETE still needs its event sent.

This is made an enum in case we find more uses for it in the future,
but it's an integer since we don't expect many of them and we can
save a few bytes that way.

The upgrade strategy is that all existing requests will have their
value set to 0, so the launchers will send a second event for any
fulfilled requests on upgrade.  This is okay because the scheduler
detects this situation, logs a warning, and otherwise takes no
further action.

Original-Change-Id: I1304ebbd3cf6ccb41fc6379a55a09d84b9f74e6c

Launcher: improve missing artifact log message

This log message only logs the missing artifact, not the upload
that actually exists.  Log both so that we know what upload to
inspect.

Original-Change-Id: Ia414ad55959bca44dfcda960c2b8242a1061d4b5

Launcher: retry loading providers on failure

If we encounter an error when loading providers, we may wait until
the next reconfiguration in order to try again, meanwhile, not
processing requests for those endpoints.

Fix this by retrying on error.

Also, catch and log errors in a provider postConfig (refreshing
quota information), but continue.

Change-Id: I853e2229baa824ee4bffcb2d7c22ea749e797ddb
Original-Change-Id: I91e28cbf93c70b432ce44939c4cafe130765d6c8