Sort nodes by request time
Whenever a node launch fails and is retried (which may happen often
due to quota), a new node record is created with a random uuid.

When we iterate over nodes to process them, we use the order that
they are returned from the cache.  This may be chronological order
if the cache has been sucessfully watching new nodes being created,
or it may be alphabetical if the launcher recently started up and
read the cache from zk.  Or it may be a mixture of the two.

This means that, at best, the order that nodes are created when there
is a backlog of nodes in the "requested" state is random.  At worst,
it's the worst case where older nodes are disadvantaged by being
replaced at the end of the list when failing.

(Note that the nodeset requests are sorted by priority and time, but
once a node is actually created with the intent to build an instance
in the cloud -- this is what is currently un-ordered.  This is not
a concern unless providers are backlogged with nodes allocated to
them in the "requested" state.)

To address this, store the time of the original nodeset request on
each of the nodes created for it (or, the current time in the case
of min-ready).  Sort nodes by request time when processing them,
so that we always attempt to drive the oldest ones first.

Existing nodes in ZK will have a request time of 0, so they will
all appear equally very old on upgrade.

The output of the nodes REST api endpoint will now also be sorted
by the same metrics.

Change-Id: I945ab9719f6904c0b0c2d0048cfcd8bbbb8acfde