Ensure node request event is sent after crash
We recently changed the order of events in the launcher so that
the node request complete event is sent after the state attribute
is set.  That is so that the scheduler can be assured that all
the attributes it needs are correct when it receives the event.
But it also means that if we crash between setting the state on
the request and sending the event, there will be no indication
to a subsequent launcher that picks up the request that the event
still needs to be sent.  It would get stuck until someone manually
dequeued a change.

To address this, we add another attribute to the request to
indicate that the event has been sent.  The intended audience is
only the launcher, not the scheduler.  Subsequent launchers can
see that a "fulfilled" request with an event_state that is not
COMPLETE still needs its event sent.

This is made an enum in case we find more uses for it in the future,
but it's an integer since we don't expect many of them and we can
save a few bytes that way.

The upgrade strategy is that all existing requests will have their
value set to 0, so the launchers will send a second event for any
fulfilled requests on upgrade.  This is okay because the scheduler
detects this situation, logs a warning, and otherwise takes no
further action.

Change-Id: I1304ebbd3cf6ccb41fc6379a55a09d84b9f74e6c