Replace getQuotaUsed with cache
Now that we store two totals for quota usage in the cache: requested
and actually used, we can replace the getQuotaUsed method with a
simple cache lookup.  This saves us from iterating over all of the
nodes every time we consider launching a node.

This will speed up the node check loop, which, in addition to simply
increasing throughput, also improves node distribution since the longer
we spend in the loop, the greater chance an opening in quota will
appear mid-loop, where we might then launch a newer node rather than
an older one.

Change-Id: I47b7b4eba405a84ab1a6ff22b786ae3ee5991049