Fix connection filter lookups in updateTenantProviders
test_launcher_connection_filter failed because this unexecpted exception
was raised during test execution:

  2025-08-15 14:50:06,789 zuul.Launcher                    ERROR    Error in main thread:
  Traceback (most recent call last):
    File "/home/zuul/src/opendev.org/zuul/zuul/zuul/launcher/server.py", line 1150, in run
      self._run()
    File "/home/zuul/src/opendev.org/zuul/zuul/tests/base.py", line 1205, in _run
      return super()._run()
             ^^^^^^^^^^^^^^
    File "/home/zuul/src/opendev.org/zuul/zuul/zuul/launcher/server.py", line 1163, in _run
      updated = self.updateTenantProviders()
                ^^^^^^^^^^^^^^^^^^^^^^^^^^^^
    File "/home/zuul/src/opendev.org/zuul/zuul/zuul/launcher/server.py", line 2224, in updateTenantProviders
      designated_launcher = self._chooseLauncherForEndpoint(
                            ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
    File "/home/zuul/src/opendev.org/zuul/zuul/zuul/launcher/server.py", line 2192, in _chooseLauncherForEndpoint
      or endpoint.connection_name in c.connection_filter}
         ^^^^^^^^^^^^^^^^^^^^^^^^
  AttributeError: 'AwsProviderEndpoint' object has no attribute 'connection_name'

I believe this occurs because endpoint's don't have connection names;
they have connections that have connection names. We fix this issue by
updating the code to use endpoint.connection.connection_name as a
result.

The reason this is not a 100% of the time failure despite the class
attribute lookup being wrong is that we only call updateTenantProviders
if the layout has been updated. This means that there is a race in the
existing test case where if the layouts have all been processed before
we add connection_filters we'll never exercise this code. In the cases
where we're still processing things in the first launcher as the second
launcher comes up the first launcher can see the second launcher in the
component registry and attempt to apply its connection_filter and then
we get this error.

Rather than simply wait until layouts are settled in the first launcher
before starting the second I've modified the test case to instead
forcefully trigger updateTenantProviders in the first launcher after
the second has been registered. This should ensure we're covering this
code path.

Change-Id: I34eb67c97ec0fa8a6fc4599df95f2aafa89de81d