DNM: add response support to MQTT reporter
We are unlikely to ever merge this change; it has some structural
problems.  Bi-directional communication via MQTT is awkward since
it involves creating non-standard, ad-hoc protocols.  This change
also requires a functioning MQTT server for tests, which we don't
normally have.

However, this change is useful for prototyping the idea of a
reporter which can wait for a result.  Reporters like this may be
used to perform synchronous out-of-band updates to git repositories
or artifact systems, but only on the condition that a merge has
happened.

If these experiments are successful, we can conisder adding a new
reporter with a more suitable protocol (such as HTTP), or perhaps
a new type of lightweight job that can run during the reporting
phase.

This change adds the ability to configure the MQTT reporter to wait
for a response during the reporting phase.

Change-Id: Ia5565cbfdbda7ed22ded2692f895833f9821ae31